🧱 refactor: extract YT_backend into dedicated yt_backend module
